ssl/letsencrypt-to-nomad-vars/requirements.txt
python-nomad>=1.5.0
absl-py
